Fix the lost event issue on Raspberry Pis
The Raspberry Pi/Qt/Chromium integration has a bug that swaps the
ordering of the lowlevel QQuickWindow and QWidgetWindow instance in a
way that is unfixable with raise()/lower() and which causes input events
to be sent to the wrong one. This fixes the issue by rerouting input
events sent to the QQuickWindow back to the QWidgetWindow. For platforms
without this bug, the rerouting never happens.